requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
h11==0.14.0
idna==3.4
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
py-notifier==0.5.0
pycparser==2.21
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1
requests==2.28.2
schedule==1.1.0
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
telebot==0.0.5
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
WinToaster==0.1.0
wsproto==1.2.0
